Update the windowed configuration for the apps
The windowed configuration GUI is basically the same for all the
platforms. It imports the auxiliar _options* templates to display
the set of objects included in HTML Kickstart.

The chosen options have been:
- A multiple selector for the platforms in usufy, phonefy and
searchfy and for domains and TLD types in mailfy and domainfy.
- A checkbox to launch the command using -w as conceived in the
last osrframework_server.py update.
- A set of checkboxes to ser the export format.
- A text field to set the file name (by default, it is usually
`profiles` to create `profiles.csv` or `profiles.json` files.