Fix issue #251: Mailfy seems to be crashing persistently
Add a timeout to the verification process so as not to be waiting
for a long time until the server sends a response. Results taking
too much time are supposed to be broken. At the same time, Win32
systems will be performing the verification operations against
validate_email libraries will others will keep using emailahoy.